backend/requirements.txt
fastapi==0.109.2
uvicorn==0.21.1
python-dotenv==1.0.0
requests==2.28.2
pydantic==1.10.15
groq
